+++ scripts/handlersManager.js
@@ -1,36 +1,42 @@
-import { ActionHandler5e } from "./actions/dnd5e/dnd5e-actions.js";
-import { ActionHandlerWfrp } from "./actions/wfrp4e/wfrp4e-actions.js";
-import { ActionHandlerPf2e } from "./actions/pf2e/pf2e-actions.js";
-import * as roll5e from "./rollHandlers/dnd5e/dnd5e-factory.js";
-import * as rollWfrp from "./rollHandlers/wfrp4e/wfrp4e-factory.js";
-import * as rollPf2e from "./rollHandlers/pf2e/pf2e-factory.js";
+import { ActionHandler5e } from './actions/dnd5e/dnd5e-actions.js';
+import { ActionHandlerWfrp } from './actions/wfrp4e/wfrp4e-actions.js';
+import { ActionHandlerPf2e } from './actions/pf2e/pf2e-actions.js';
+import { ActionHandlerDw } from './actions/dungeonworld/dw-actions.js';
+import * as roll5e from './rollHandlers/dnd5e/dnd5e-factory.js';
+import * as rollWfrp from './rollHandlers/wfrp4e/wfrp4e-factory.js';
+import * as rollPf2e from './rollHandlers/pf2e/pf2e-factory.js';
+import * as rollDw from './rollHandlers/dungeonworld/dw-factory.js';
 
 export class HandlersManager {
     // Currently only planning for one kind of action handler for each system
     static getActionHandler(system) {
         switch (system) {
-            case "dnd5e":
+            case 'dnd5e':
                 return new ActionHandler5e();
-            case "pf2e":
+            case 'pf2e':
                 return new ActionHandlerPf2e();
-            case "wfrp4e":
+            case 'wfrp4e':
                 return new ActionHandlerWfrp();
+            case 'dungeonworld':
+                return new ActionHandlerDw();
         }
-        throw new Error("System not supported by Token Action HUD");
+        throw new Error('System not supported by Token Action HUD');
     }
 
     // Possibility for several types of rollers (e.g. BetterRolls, MinorQOL for DND5e),
     // so pass off to a RollHandler factory
     static getRollHandler(system, handlerId) {
         switch (system) {
-            case "dnd5e":
+            case 'dnd5e':
                 return roll5e.getRollHandler(handlerId)
-            case "pf2e":
+            case 'pf2e':
                 return rollPf2e.getRollHandler(handlerId);
-            case "wfrp4e":
+            case 'wfrp4e':
                 return rollWfrp.getRollHandler(handlerId);
+            case 'dungeonworld':
+                return rollDw.getRollHandler(handlerId);
         }
-        throw new Error("System not supported by Token Action HUD");
+        throw new Error('System not supported by Token Action HUD');
     }
 
     // Not yet implemented.
@@ -38,16 +44,19 @@ export class HandlersManager {
         let choices;
 
         switch (system) {
-            case "dnd5e":
-                choices = {"core": "Core 5e"};
-                this.testForModule(choices, "betterrolls5e");
-                this.testForModule(choices, "minor-qol");
+            case 'dnd5e':
+                choices = {'core': 'Core 5e'};
+                this.testForModule(choices, 'betterrolls5e');
+                this.testForModule(choices, 'minor-qol');
                 break;
-            case "pf2e":
-                choices = {"core": "Core PF2E"};
+            case 'pf2e':
+                choices = {'core': 'Core PF2E'};
                 break;
-            case "wfrp4e":
-                choices = {"core": "Core Wfrp"};
+            case 'wfrp4e':
+                choices = {'core': 'Core Wfrp'};
+                break;
+            case 'dungeonworld':
+                choices = {'core': 'Core DungeonWorld'};
                 break;
         }